+++ fact/src/main/java/com/team23/fact/domain/models/CategoryModel.kt
@@ -2,5 +2,7 @@ package com.team23.fact.domain.models
 
 data class CategoryModel(
     val nameEn: String,
-    val nameFr: String
+    val nameFr: String,
+    val shortNameEn: String,
+    val shortNameFr: String,
 )