+++ fact/src/main/java/com/team23/fact/domain/usecases/GetCategoryShortNameUseCase.kt
@@ -4,15 +4,15 @@ import androidx.compose.ui.text.intl.Locale
 import com.team23.fact.domain.repositories.CategoryRepository
 import javax.inject.Inject
 
-class GetCategoryUseCase @Inject constructor(
+class GetCategoryShortNameUseCase @Inject constructor(
     private val categoryRepository: CategoryRepository
 ) {
-    suspend fun execute(code: String?) = code?.let {
+    suspend operator fun invoke(code: String?): String = code?.let {
         categoryRepository.getCategoryNameByCode(code).let { category ->
             if (Locale.current.language == "fr") {
-                category.nameFr
+                category.shortNameFr
             } else {
-                category.nameEn
+                category.shortNameEn
             }
         }
     } ?: ""