+++ fact/src/main/java/com/team23/fact/presentation/viewobjects/CategoryVO.kt
@@ -1,0 +1,7 @@
+package com.team23.fact.presentation.viewobjects
+
+data class CategoryVO(
+    val code: String?,
+    val name: String,
+    val shortName: String
+)