+++ fact/src/main/java/com/team23/fact/presentation/viewobjects/FactDetailVO.kt
@@ -5,8 +5,7 @@ import androidx.compose.ui.graphics.ImageBitmap
 data class FactDetailVO(
     val id: String = "",
     val title: String = "",
-    val category: String = "",
-    val codeCategory: String? = null,
+    val category: CategoryVO,
     val imageUrl: String? = null,
     val imageBitmap: ImageBitmap? = null,
     val description: String = "",