+++ home/src/main/java/com/team23/home/domain/models/CategoryModel.kt
@@ -3,5 +3,7 @@ package com.team23.home.domain.models
 data class CategoryModel(
     val code: String,
     val titleEn: String,
-    val titleFr: String
+    val titleFr: String,
+    val shortTitleEn: String,
+    val shortTitleFr: String
 )