+++ home/src/main/java/com/team23/home/presentation/viewobjects/CategoryVO.kt
@@ -2,5 +2,6 @@ package com.team23.home.presentation.viewobjects
 
 data class CategoryVO(
     val code: String,
-    val title: String
+    val title: String,
+    val shortTitle: String,
 )